Lots of cleanup.
I was previously coding in ipynb notebooks, with the setup so that each ipynb file had a corresponding .py file. This resulted in lots of extra gunk. Cleaned up the .py files for actual jupyter notebooks, and the .ipynb files for actual python libraries.

Updated code for Tensorflow versions of Curveball and LMA.